rpc refactor for accounts microservice
