+++ accounts/src/app.js
@@ -1,7 +1,6 @@
 import express from 'express';
 import cors from 'cors';
 import bodyParser from 'body-parser';
-import account from './routes/account';
 import accounts from './routes/accounts';
 import { formatResponse, formatError, errorHandler } from './middlewares';
 
@@ -12,8 +11,7 @@ app.use(cors());
 app.use(bodyParser.json());
 
 // Routes
-app.use('/account', account);
-app.use('/accounts', accounts);
+app.use('/', accounts);
 
 app.use(formatResponse);
